python-service/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
beautifulsoup4==4.12.3
lxml==5.1.0

# Text Processing
nltk==3.8.1
unidecode==1.3.8

# AI/ML
openai>=1.12.0
tiktoken==0.5.2
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
transformers==4.37.2

# Vector Storage
redis==5.0.1
numpy==1.26.3

# Database
mysql-connector-python==8.3.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

aiohttp==3.9.1
beautifulsoup4==4.12.3
lxml==5.1.0